Updated types to be more explicit.

+++ Relay.cpp
@@ -10,35 +10,35 @@ RELAY_SIZE *Relay::getSize(void) {
 	return &this->_size;
 }
 
-Relay &Relay::setPin(unsigned int pin, unsigned int relay) {
+Relay &Relay::setPin(uint8_t pin, uint8_t relay) {
 	this->_pins[relay - 1] = pin;
 	pinMode(pin, OUTPUT);
 	digitalWrite(pin, HIGH);
 	return *this;
 }
 
-int Relay::getRelayByPin(unsigned int pin) {
-	for (unsigned short int i = 0; i < *this->getSize(); i++) {
+uint8_t Relay::getRelayByPin(uint8_t pin) {
+	for (uint8_t i = 0; i < *this->getSize(); i++) {
 		if (this->_pins[i] == pin) { return i + 1; }
 	}
 	return 0;
 }
 
-int Relay::getPinByRelay(unsigned int relay) {
+uint8_t Relay::getPinByRelay(uint8_t relay) {
 	return &this->_pins[relay - 1];
 }
 
-Relay &Relay::on(unsigned int relay) {
+Relay &Relay::on(uint8_t relay) {
 	this->_futurePinState[relay - 1] = RELAY_ON;
 	return *this;
 }
 
-Relay &Relay::off(unsigned int relay) {
+Relay &Relay::off(uint8_t relay) {
 	this->_futurePinState[relay - 1] = RELAY_OFF;
 	return *this;
 }
 
-Relay &Relay::toggle(unsigned int relay) {
+Relay &Relay::toggle(uint8_t relay) {
 	if (*this->getState(relay) == RELAY_OFF) {
 		return this->on(relay);
 	} else {
@@ -47,10 +47,10 @@ Relay &Relay::toggle(unsigned int relay) {
 }
 
 Relay &Relay::commit(void) {
-	for (unsigned short int i = 0; i < *this->getSize(); i++) {
+	for (uint8_t i = 0; i < *this->getSize(); i++) {
 		RELAY_STATE &c = this->_currentPinState[i];
 		RELAY_STATE &f = this->_futurePinState[i];
-		unsigned long *l = &this->_lastToggle[i];
+		uint32_t *l = &this->_lastToggle[i];
 
 		if (c != f && Avail::millis(&this->_safety, l)) {
 			c = f;
@@ -62,10 +62,10 @@ Relay &Relay::commit(void) {
 	return *this;
 }
 
-RELAY_STATE *Relay::getState(unsigned int relay) {
+RELAY_STATE *Relay::getState(uint8_t relay) {
 	return &this->_currentPinState[relay - 1];
 }
 
-Relay &Relay::setSafety(unsigned long safety) {
+Relay &Relay::setSafety(uint32_t safety) {
 	this->_safety = max(safety, 250);
 }
